src/multi_agent_code_analyzer/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
neo4j>=4.4.0
redis>=4.0.0
pymilvus>=2.0.0
minio>=7.0.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0
asyncio>=3.4.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
prometheus-client>=0.11.0
numpy>=1.24.0
scikit-learn>=1.3.2 